refactor: edit response messages
