+++ src/controllers/turn.ts
@@ -1,6 +1,6 @@
 import { db } from '../db';
-import { ATTACK_STATUS, MESSAGE_TYPES } from '../types/enums';
-import { IMessage } from '../types/interfaces';
+import { ATTACK_STATUS } from '../types/enums';
+import { turnResponse } from '../utils/response';
 
 export const turn = (gameId: number, status?: string) => {
   const { findGame, sockets } = db;
@@ -11,13 +11,8 @@ export const turn = (gameId: number, status?: string) => {
     status === ATTACK_STATUS.MISS ? (currentPlayer === 0 ? 1 : 0) : currentPlayer;
 
   game.players.forEach((player) => {
-    const message: IMessage = {
-      type: MESSAGE_TYPES.TURN,
-      data: JSON.stringify({
-        currentPlayer: game.players[game.currentPlayer].index,
-      }),
-      id: 0,
-    };
-    sockets[player.index].send(JSON.stringify(message));
+    const currentPlayer = game.players[game.currentPlayer].index;
+    const message = turnResponse(currentPlayer);
+    sockets[player.index].send(message);
   });
 };
