+++ src/controllers/updateRoom.ts
@@ -1,16 +1,8 @@
 import { db } from '../db';
-import { MESSAGE_TYPES } from '../types/enums';
+import { updateRoomsResponse } from '../utils/response';
 
 export const updateRooms = () => {
   const { sockets, rooms } = db;
 
-  Object.keys(sockets).forEach((key) =>
-    sockets[key].send(
-      JSON.stringify({
-        type: MESSAGE_TYPES.UPDATE_ROOM,
-        data: JSON.stringify(rooms),
-        id: 0,
-      })
-    )
-  );
+  Object.keys(sockets).forEach((key) => sockets[key].send(updateRoomsResponse(rooms)));
 };
